+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#4164](https://github.com/open-telemetry/opentelemetry-python/pull/4346))
 - Tolerates exceptions when loading resource detectors via `OTEL_EXPERIMENTAL_RESOURCE_DETECTORS`
   ([#4373](https://github.com/open-telemetry/opentelemetry-python/pull/4373))
+- opentelemetry-sdk: fix OTLP exporting of Histograms with explicit buckets advisory
+  ([#4434](https://github.com/open-telemetry/opentelemetry-python/pull/4434))
+- opentelemetry-exporter-otlp-proto-grpc: better dependency version range for Python 3.13
+  ([#4444](https://github.com/open-telemetry/opentelemetry-python/pull/4444))
+- opentelemetry-exporter-opencensus: better dependency version range for Python 3.13
+  ([#4444](https://github.com/open-telemetry/opentelemetry-python/pull/4444))
 
 ## Version 1.30.0/0.51b0 (2025-02-03)